baseline_code/requirements.txt
mango-agents==0.4.0
pandas==1.5.3
pytest==7.2.2
pytest-asyncio==0.21.0
statsmodels==0.13.5
